Add the reset script
This script will take input information, url and password, log into a BGW210 modem/router, and reset it. This is intended for if a user has a problem with the modem disconnecting regularly from the internet and requires a reset to reconnect. This probably shouldn't be a permanent fix, and a call to AT&T is more likely to be a better fix, but sharing in case someone finds it useful.